refactor: split server routes into modules
